src/core/yolo/requirements.txt
# Usage: pip install -r requirements.txt

# Base ----------------------------------------
matplotlib
numpy
opencv-python
Pillow
PyYAML
requests
scipy
torch!=1.12.0
torchvision!=0.13.0
tqdm
protobuf

# Logging -------------------------------------
tensorboard
# wandb

# Plotting ------------------------------------
pandas
seaborn

# Export --------------------------------------
# coremltools>=4.1  # CoreML export
# onnx>=1.9.0  # ONNX export
# onnx-simplifier>=0.3.6  # ONNX simplifier
# scikit-learn==0.19.2  # CoreML quantization
# tensorflow>=2.4.1  # TFLite export
# tensorflowjs>=3.9.0  # TF.js export
# openvino-dev  # OpenVINO export

# Extras --------------------------------------
ipython  # interactive notebook
psutil  # system utilization
thop  # FLOPs computation
# albumentations>=1.0.3
# pycocotools>=2.0  # COCO mAP
# roboflow
pynvml
mlflow
supervision
prettytable
onnx-simplifier
tritonclient[all]
fastapi
uvicorn